:tada: feat: add fetch
